pyproject.toml
[project]
name = "rts-ai-platform"
version = "0.1.0"
description = "AI Native RTS Game Development Platform — 2D headless trainable replayable batch-simulatable RTS combat sandbox"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "protobuf>=5.0",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "numpy>=1.26",
    "pydantic>=2.5",
]

[project.optional-dependencies]
agents = [
    "agentscope>=0.1",
]
train = [
    "torch>=2.2",
    "trl>=0.12",
]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "mypy-protobuf>=3.6",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["simcore/proto_out"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"simcore/grpc_server.py" = ["N802"]  # protobuf method names must be PascalCase

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "grpc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "agentscope.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -n 4"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]